Update Basic_Calculator.py
No Changes. This commit was made because I made a small mistake in the commit message in the earlier commit. It should be, "Added a few spaces and newlines along with another while loop so that it can work continuously and a break statement to exit the loop if the user decides not to run the program anymore."